requirements.txt
fastapi
uvicorn[standard]
requests
streamlit
sentence-transformers
faiss-cpu
numpy
pyyaml
